+++ strawberry/experimental/pydantic/object_type.py
@@ -15,7 +15,6 @@ from typing import (
     Type,
     cast,
 )
-from pydantic import BaseModel
 
 from strawberry.annotation import StrawberryAnnotation
 from strawberry.auto import StrawberryAuto
@@ -37,15 +36,10 @@ from strawberry.experimental.pydantic.utils import (
     get_private_fields,
 )
 from strawberry.field import StrawberryField
-from strawberry.object_type import _get_interfaces, _process_type, _wrap_dataclass
+from strawberry.object_type import _process_type, _wrap_dataclass
 from strawberry.types.type_resolver import _get_fields
 from strawberry.utils.dataclasses import add_custom_init_fn
 
-from strawberry.utils.str_converters import to_camel_case
-
-from strawberry.types.types import StrawberryObjectDefinition
-
-from strawberry.utils.deprecations import DEPRECATION_MESSAGES, DeprecatedDescriptor
 
 if TYPE_CHECKING:
     from graphql import GraphQLResolveInfo
@@ -121,161 +115,6 @@ if TYPE_CHECKING:
         PydanticModel,
         StrawberryTypeFromPydantic,
     )
-
-
-def _get_strawberry_fields_from_pydantic(
-    model: Type[BaseModel], is_input: bool, use_pydantic_alias: bool
-) -> List[StrawberryField]:
-    """Get all the strawberry fields off a pydantic BaseModel cls
-
-    This function returns a list of StrawberryFields (one for each field item), while
-    also paying attention the name and typing of the field.
-
-
-    Type #1:
-        A pure pydantic field. Will not have a StrawberryField; one will need to
-        be created in this function. Type annotation is required.
-    """
-    fields: list[StrawberryField] = []
-
-    # BaseModel already has fields, so we need to get them from there
-    model_fields: Dict[str, CompatModelField] = get_model_fields(model)
-    for name, field in model_fields.items():
-        converted_type = replace_types_recursively(field.outer_type_, is_input=is_input)
-        if field.allow_none:
-            converted_type = Optional[converted_type]
-        fields.append(
-            StrawberryField(
-                python_name=name,
-                graphql_name=field.alias if use_pydantic_alias else None,
-                # always unset because we use default_factory instead
-                default=dataclasses.MISSING,
-                default_factory=get_default_factory_for_field(field),
-                type_annotation=StrawberryAnnotation.from_annotation(converted_type),
-                description=field.description,
-                deprecation_reason=None,
-                permission_classes=[],
-                directives=[],
-                metadata={},
-            )
-        )
-
-    return fields
-
-
-def first_class_process_base_model(
-    model: Type[BaseModel],
-    *,
-    name: Optional[str] = None,
-    is_input: bool = False,
-    is_interface: bool = False,
-    description: Optional[str] = None,
-    directives: Optional[Sequence[object]] = (),
-    extend: bool = False,
-    use_pydantic_alias: bool = True,
-):
-    name = name or to_camel_case(model.__name__)
-
-    interfaces = _get_interfaces(model)
-    fields = _get_strawberry_fields_from_pydantic(
-        model, is_input=is_input, use_pydantic_alias=use_pydantic_alias
-    )
-    is_type_of = getattr(model, "is_type_of", None)
-    resolve_type = getattr(model, "resolve_type", None)
-
-    model.__strawberry_definition__ = StrawberryObjectDefinition(
-        name=name,
-        is_input=is_input,
-        is_interface=is_interface,
-        interfaces=interfaces,
-        description=description,
-        directives=directives,
-        origin=model,
-        extend=extend,
-        fields=fields,
-        is_type_of=is_type_of,
-        resolve_type=resolve_type,
-    )
-    # TODO: remove when deprecating _type_definition
-    DeprecatedDescriptor(
-        DEPRECATION_MESSAGES._TYPE_DEFINITION,
-        model.__strawberry_definition__,
-        "_type_definition",
-    ).inject(model)
-
-    return model
-
-
-def register_first_class(
-    model: Type[PydanticModel],
-    *,
-    name: Optional[str] = None,
-    is_input: bool = False,
-    is_interface: bool = False,
-    description: Optional[str] = None,
-    use_pydantic_alias: bool = True,
-) -> Type[PydanticModel]:
-    """Registers a pydantic model as a first class strawberry type.
-
-    This is useful if you want to use a pydantic model as a type in a field
-    without having to create a separate strawberry type.
-
-    Example:
-
-        class User(BaseModel):
-            id: int
-            name: str
-
-        register_first_class(User)
-
-        @strawberry.type
-        class Query:
-            @strawberry.field
-            def user(self) -> User:
-                return User(id=1, name="Patrick")
-
-
-    """
-
-    first_class_process_base_model(
-        model,
-        name=name,
-        is_input=is_input,
-        is_interface=is_interface,
-        description=description,
-        use_pydantic_alias=use_pydantic_alias,
-    )
-
-    if is_input:
-        # TODO: Probably should check if the name clashes with an existing type?
-        model._strawberry_input_type = model  # type: ignore
-    else:
-        model._strawberry_type = model  # type: ignore
-
-    return model
-
-
-def first_class_type(
-    name: Optional[str] = None,
-    is_input: bool = False,
-    is_interface: bool = False,
-    description: Optional[str] = None,
-    use_pydantic_alias: bool = True,
-) -> Callable[[Type[PydanticModel]], Type[PydanticModel]]:
-    """A decorator to make a pydantic class work on strawberry without creating
-    a separate strawberry type."""
-
-    def wrap(model: Type[PydanticModel]) -> Type[PydanticModel]:
-        return register_first_class(
-            model,
-            name=name,
-            is_input=is_input,
-            is_interface=is_interface,
-            description=description,
-            use_pydantic_alias=use_pydantic_alias,
-        )
-
-    return wrap
 
 
 def type(
